# 14 simplify determination of formatting functions