```console
$ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                         │    a.txt     │                b.txt                │
                         │    sec/op    │   sec/op     vs base                │
NewDateTime/fa-IR-12       211.05n ± 2%   67.87n ± 1%  -67.84% (p=0.000 n=10)
NewDateTime/lv-LV-12        263.7n ± 1%   103.8n ± 2%  -60.62% (p=0.000 n=10)
NewDateTime/dz-BT-12        265.0n ± 1%   106.4n ± 1%  -59.87% (p=0.000 n=10)
DateTime_Format/fa-IR-12    859.9n ± 1%   595.9n ± 0%  -30.70% (p=0.000 n=10)
DateTime_Format/lv-LV-12    174.6n ± 1%   147.7n ± 1%  -15.41% (p=0.000 n=10)
DateTime_Format/dz-BT-12    458.1n ± 1%   253.5n ± 1%  -44.67% (p=0.000 n=10)
geomean                     317.0n        159.9n       -49.55%

                         │    a.txt     │               b.txt                │
                         │     B/op     │    B/op     vs base                │
NewDateTime/fa-IR-12       269.000 ± 0%   5.000 ± 0%  -98.14% (p=0.000 n=10)
NewDateTime/lv-LV-12       221.000 ± 0%   5.000 ± 0%  -97.74% (p=0.000 n=10)
NewDateTime/dz-BT-12       221.000 ± 0%   5.000 ± 0%  -97.74% (p=0.000 n=10)
DateTime_Format/fa-IR-12    104.00 ± 0%   72.00 ± 0%  -30.77% (p=0.000 n=10)
DateTime_Format/lv-LV-12     32.00 ± 0%   32.00 ± 0%        ~ (p=1.000 n=10)
DateTime_Format/dz-BT-12    144.00 ± 0%   64.00 ± 0%  -55.56% (p=0.000 n=10)
geomean                      135.9        16.25       -88.04%

                         │    a.txt    │                b.txt                 │
                         │  allocs/op  │ allocs/op   vs base                  │
NewDateTime/fa-IR-12        5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=10)
NewDateTime/lv-LV-12        5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=10)
NewDateTime/dz-BT-12        5.000 ± 0%   1.000 ± 0%  -80.00% (p=0.000 n=10)
DateTime_Format/fa-IR-12   23.000 ± 0%   9.000 ± 0%  -60.87% (p=0.000 n=10)
DateTime_Format/lv-LV-12    4.000 ± 0%   4.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   24.000 ± 0%   8.000 ± 0%  -66.67% (p=0.000 n=10)
geomean                     8.069        2.570       -68.15%
¹ all samples are equal
```